requirements.txt
# requirements.txt

langchain
streamlit
weaviate-haystack
trulens-eval
